Add .wasm file symlinks
We want to re-use examples/testapp/testapp.wasm to avoid
having to build the sample app several times.

Emscripten loads the wasm file relative to the html
file, and not relative to the app.js file.

Create a symlink in each of the example directories
which points to the (one) testapp.wasm.