Add dynamic linking example / debug helper
This is a work in progress. This example shows how to
load a basic Qt Quick application using shared libraries
with pre-loaded Qt plugins, and also how to load a plugin
on demand.

See documentation at the doc snapshots for how to build
Qt and this example

https://doc-snapshots.qt.io/qt6-dev/wasm.html#shared-libraries-and-dynamic-linking-developer-preview

Note that this checkin contains generated plugin preload lists
(qt_plugins.json and qt_qml_imports.json). This means that
running the scripts to generate them may be skipped.